refactoring unit tests
